add odom_data instead of pose_data (#12)
* add odom data and update odometry_publisher

* update odometry_subscriber

* remoev pose data

* add comment for odom data

* format code

+++ lidar_localization/src/matching_node.cpp
@@ -146,7 +146,7 @@ bool MatchingNode::update_matching()
   if (!matching_->has_inited()) {
     // global initialization
     if (matching_->set_init_pose_by_scan_context(current_lidar_data_)) {
-      Eigen::Matrix4f init_pose = matching_->get_init_pose();
+      Eigen::Matrix4d init_pose = matching_->get_init_pose().cast<double>();
       // evaluate deviation from GNSS/IMU:
       float deviation =
         (init_pose.block<3, 1>(0, 3) - current_gnss_data_.pose.block<3, 1>(0, 3)).norm();
@@ -154,7 +154,7 @@ bool MatchingNode::update_matching()
                 << deviation << std::endl;
     } else {
       // if failed, fall back to GNSS/IMU init:
-      matching_->set_init_pose_by_gnss(current_gnss_data_.pose);
+      matching_->set_init_pose_by_gnss(current_gnss_data_.pose.cast<float>());
       std::cout << "Scan Context Localization Init Failed. Fallback to GNSS/IMU." << std::endl;
     }
   }
@@ -163,7 +163,7 @@ bool MatchingNode::update_matching()
 
 bool MatchingNode::publish_data()
 {
-  lidar_odom_pub_->publish(lidar_odometry_, current_lidar_data_.time);
+  lidar_odom_pub_->publish(lidar_odometry_.cast<double>(), current_lidar_data_.time);
   if (publish_tf_) {
     auto msg =
       localization_common::to_transform_stamped_msg(lidar_odometry_, current_lidar_data_.time);
